requirements.txt
# Smart Study Assistant - Python Dependencies

# Google AI and Agent Development Kit
google-adk>=0.3.0
google-generativeai>=0.8.0

# Environment management
python-dotenv>=1.0.0

# Additional utilities (optional but recommended)
# Uncomment if needed:
# requests>=2.31.0
# numpy>=1.24.0
# pandas>=2.0.0
